refactor(frontend): move network config state, save path and audit into useNetworkConfig hook
